Revert "home tilt is for mcos"
This reverts commit b8f9633f684257e38d03e08d45cf16a1dded7326.

+++ Tiltfile
@@ -2,26 +2,27 @@ analytics_settings(False)
 
 load("ext://uibutton", "cmd_button", "location")
 
-# vault
-local_resource("vault",
-    serve_cmd=[
-        "bash", "-c",
-        """
-            eval "$(direnv hook bash)"
-            direnv allow
-            _direnv_hook
-            exec this-vault-start -log-level debug
-        """
-    ]
-)
+if "-darwin" in os.getenv("system"):
+    # vault
+    local_resource("vault",
+        serve_cmd=[
+            "bash", "-c",
+            """
+                eval "$(direnv hook bash)"
+                direnv allow
+                _direnv_hook
+                exec this-vault-start -log-level debug
+            """
+        ]
+    )
 
-# nix-cache
-local_resource("nix-cache",
-    serve_cmd=[
-        "bash", "-c",
-        """
-            docker rm -f nix-cache || true
-            exec docker run --name nix-cache --rm -v nix-cache:/usr/share/nginx/html:ro -p 5001:80 nginx
-        """
-    ]
-)
+    # nix-cache
+    local_resource("nix-cache",
+        serve_cmd=[
+            "bash", "-c",
+            """
+                docker rm -f nix-cache || true
+                exec docker run --name nix-cache --rm -v nix-cache:/usr/share/nginx/html:ro -p 5001:80 nginx
+            """
+        ]
+    )
